change initial ability use limit settings' step to 0.05

+++ Roles/Crewmate/Bloodhound.cs
@@ -64,7 +64,7 @@ public class Bloodhound : RoleBase
     public override void Add(byte playerId)
     {
         PlayerIdList.Add(playerId);
-        playerId.SetAbilityUseLimit(UseLimitOpt.GetInt());
+        playerId.SetAbilityUseLimit(UseLimitOpt.GetFloat());
         BloodhoundTargets = [];
         BloodhoundId = playerId;
     }
