+++ Roles/Crewmate/Divinator.cs
@@ -71,7 +71,7 @@ public class Divinator : RoleBase
     public override void Add(byte playerId)
     {
         PlayerIdList.Add(playerId);
-        playerId.SetAbilityUseLimit(CheckLimitOpt.GetInt());
+        playerId.SetAbilityUseLimit(CheckLimitOpt.GetFloat());
 
         LateTask.New(() =>
         {
