+++ Roles/Crewmate/Doormaster.cs
@@ -45,7 +45,7 @@ public class Doormaster : RoleBase
     public override void Add(byte playerId)
     {
         PlayerIdList.Add(playerId);
-        playerId.SetAbilityUseLimit(UseLimitOpt.GetInt());
+        playerId.SetAbilityUseLimit(UseLimitOpt.GetFloat());
     }
 
     public override void Remove(byte playerId)
