+++ Roles/Crewmate/Tether.cs
@@ -54,7 +54,7 @@ public class Tether : RoleBase
     public override void Add(byte playerId)
     {
         PlayerIdList.Add(playerId);
-        playerId.SetAbilityUseLimit(UseLimitOpt.GetInt());
+        playerId.SetAbilityUseLimit(UseLimitOpt.GetFloat());
         Target = byte.MaxValue;
         TetherId = playerId;
     }
